Refactor models and schema to match python naming
